% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm_withrcpp.R
\name{my_lm_withrcpp}
\alias{my_lm_withrcpp}
\title{Linear Regression with Rcpp}
\usage{
my_lm2(X, y)
}
\arguments{
\item{X}{A matrix of predictor variables.}

\item{y}{A numeric vector of response variables.}
}
\value{
A list with components including coefficients, fitted values, residuals, R-squared,
        adjusted R-squared, standard error, F-statistic, degrees of freedom, and residual summary.
}
\description{
Using Rcpp to improve my_lm function
}
\examples{
data(mtcars)
X <- as.matrix(mtcars[, c("wt", "hp")])
y <- mtcars$mpg
result <- my_lm2(X, y)
print(result)
}
