% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals4.R
\name{residuals4}
\alias{residuals4}
\title{Calculate Residuals}
\usage{
residuals4(model)
}
\arguments{
\item{model}{A linear model object created using lm().}
}
\value{
A list containing raw residuals, standardized residuals, internally studentized residuals,
        and externally studentized residuals.
}
\description{
This function calculates raw residuals, standardized residuals, internally studentized residuals,
and externally studentized residuals for a fitted linear model.
}
\examples{
data(mydata)
model <- lm(y ~ X, data = mydata)
residuals <- residuals4(model)
print(residuals)
}
